Add some options that controls AlphaPhase (commit ce83a17)
A new options has been added in the spec file of AlphaImpute: LargeDatasets.
This option speed-ups running time of AlphaPhase. If set to 'Yes', this
option expects other two more parameters that controls the Subset size and
the number of iterations of AlphaPhase (commit: ce83a17). The other
parameters needed for this speed-up are set with the default parameters.

For mor information about this parameters, look into the section
1.2.1.1. AlphaPhaseSpec.txt of the AlphaPhase's User manual.