Bugfix: Missing values after running HMM
Some snps were missing after running the HMM algorithms which is
clearly wrong, as the HMM imputes all snps and alleles.

Use the proper variable from the HMM to impute alleles based on
probabilities.

Fix the index of the genotyped individuals instead of the index
of associated with all the individuals in the pedigree when
imputing alleles.

Not check by alleles inconsistencies.