Debug: Write probabilities fails when nAnisP != nAnisG
When the number of animals in the pedigree file is different than
the number of animals in the genotype file, AlphaImpute crashed
while writing out the genotype probabilities with the hmm option.
This was because in the hmm only the genotyped animals are considered.

Now, AlphaImpute writes out the genotype probabilities of the
animals in the genotype file if the HMM is used.