Bugfix: AlphaImpute crash before writing out results from HMM
AlphaImpute crashed because the number of individuals defined as
genotyped was less than the number of individuals in the genotype
file. The reason for this is the removal of snps due to mendelian
inconsistencies.

The solution for this bug was, to warn the user if this happens and
instead of using the number of individuals genotyped during the HMM,
using the number of individuals of the genotype file.

This potentially introduce another bug which is crashing before writing
results from HMM if there are animals in the genotype file that are
not in the pedigree file.