Feature: Support prior allele frequencies
Prior Allele frequencies needed to create haplotypes template in the HMM
can be provided by the user in the form of a list of real numbers in a
file. The file that contains the allele frequencies should be specified
in the option 'PriorAlleleFrequencies'

If the file with the allele frequencies are not provided, allele
frequencies are computed with population information.